pyproject.toml
[project]
name = "fastapi-practice"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.24.0.post1",
    "SQLAlchemy>=2.0.23",
    "aiomysql>=0.2.0",
    "pyjwt>=2.8.0",
    "polars>=0.20.16",
    "isort>=5.13.2",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.1.5",
    "pytest-asyncio>=0.23.2",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.1",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "pytest-check>=2.2.2",
    "mypy>=1.8.0",
    "types-pytz>=2024.1.0.20240203",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.rye.scripts]
lint = { chain = ["lint:black", "lint:ruff", "lint:mypy", "lint:isort" ] }
"lint:black" = "black ."
"lint:ruff" = "ruff check . --fix"
"lint:mypy" = "mypy ./src" # testディレクトリは無視
"lint:isort" = "isort ."

[tool.pytest.ini_options]
addopts = ["-p no:warnings", "-p no:logging", "--last-failed", "-s"]